Overhaul of Scratch.java
Implemented ScratchImage, a class which holds 2 GreenfootImages. One image is immutable, and thus will remain the same throughout the life of the scratchImage. The second image is generated by applying effects to the base, such as rotation, ghost, and size.Costumes now store a ScratchImage instead of 2 greeenfoot images.  Because everything now has access to both a modified and unmodified version of every costume, many functions have been significantly shortened. ScratchImages are always resized so that they are centered exactly in an image large enough that they can rotate without being cut off. With some minor modification, the center could be changed to allow un-centered rotation.