Implemented remain graphics effects
The performance on these is relatively poor compared to scratch, but scratch uses Pixel Bender Toolkit, which is very efficient at filtering images by splitting the load between the gpu and all cpu cores, so it would be difficult to reach that level with java.
I have added IntStream parallelism which has sped it up by a factor of 2-3.
Also the images are only updated once per frame, rather than every time a value is changed.