Added a quadtree for collision detection Resolve #146
Also fixed a few bugs, and improved where say actors place themselves in relation to the sprite.
Note that isTouchingColor currently does NOT use the tree. This means that it is very slow compared to isTouching. Because the tree only works for boolean values (transparency) and not numeric values (color) , it would be not be possible to use the tree the same way. There may still be some optimization possible with it, however.