Added s2g-experimental, which uses an Abstract Syntax Tree
See the file header for more information.
It is currently able to generate code for the most blocks. Notable exceptions are Variables, Lists, Graphics effects, and math functions (sqrt, exp, etc).
It also does not currently copy things over such as images or scratch.java, and does not generate a project.greenfoot file. Fixing this should be a matter of copying the code from s2g.